radio-bcm2048: Fix region selection
This actually fixes region selection for BCM2048 FM receiver. To select
the japanese FM-band an additional bit in FM_CTRL register needs to be
set. This might not sound so important but it enables at least me to
listen to some 'very interesting' radio transmission below normal
FM-band.

Removed "Japan wide band" region since this is impossible to do just
like that. Additionally it's now possible to go back to non-Japanese
regions without having to reload the module.

Signed-off-by: Pali Rohár <pali.rohar@gmail.com>